+++ src/Exceptions/UnsupportedOperationException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Samsara\Exceptions;
+
+/**
+ * Exception for operations not supported by the Samsara API.
+ *
+ * @author Erik Galloway <erik@erikgall.com>
+ */
+class UnsupportedOperationException extends SamsaraException
+{
+    //
+}